+++ spec/unit/resolver_spec.rb
@@ -378,9 +378,28 @@ module Pod
         e.message.should.match(/`AFNetworking \(= 999\.999\.999\)` required by `Podfile`/)
         e.message.should.match(/None of your spec sources contain a spec satisfying the dependency: `AFNetworking \(= 999\.999\.999\)`./)
         e.message.should.match(/You have either:/)
-        e.message.should.match(/ * out-of-date source repos which you can update with `pod repo update`/)
+        e.message.should.match(/ * out-of-date source repos which you can update with `pod repo update` or with `pod install --repo-update`./)
+        e.message.should.match(/ * mistyped the name or version./)
         e.message.should.match(/ * not added the source repo that hosts the Podspec to your Podfile./)
-        e.message.should.match(/ * out-of-date source repos which you can update with `pod repo update`/)
+        e.message.should.match(/Note: as of CocoaPods 1.0, `pod repo update` does not happen on `pod install` by default./)
+        e.exit_status.should.equal(31)
+      end
+
+      it 'raises if repo are updated and no such version of a dependency exists' do
+        podfile = Podfile.new do
+          platform :ios
+          pod 'AFNetworking', '999.999.999'
+        end
+        resolver = Resolver.new(config.sandbox, podfile, empty_graph, config.sources_manager.all)
+        resolver.specs_updated = true
+        e = lambda { resolver.resolve }.should.raise NoSpecFoundError
+        e.message.should.match(/Unable to satisfy the following requirements/)
+        e.message.should.match(/`AFNetworking \(= 999\.999\.999\)` required by `Podfile`/)
+        e.message.should.match(/None of your spec sources contain a spec satisfying the dependency: `AFNetworking \(= 999\.999\.999\)`./)
+        e.message.should.match(/You have either:/)
+        e.message.should.not.match(/ * out-of-date source repos which you can update with `pod repo update` or with `pod install --repo-update`./)
+        e.message.should.match(/ * mistyped the name or version./)
+        e.message.should.match(/ * not added the source repo that hosts the Podspec to your Podfile./)
         e.message.should.match(/Note: as of CocoaPods 1.0, `pod repo update` does not happen on `pod install` by default./)
         e.exit_status.should.equal(31)
       end
@@ -400,9 +419,9 @@ module Pod
         e.message.should.match(/None of your spec sources contain a spec satisfying the dependencies:/)
         e.message.should.match(/`AFNetworking \(= 3.0.1\), AFNetworking \(= 1.4\)`/)
         e.message.should.match(/You have either:/)
-        e.message.should.match(/ * out-of-date source repos which you can update with `pod repo update`/)
+        e.message.should.match(/ * out-of-date source repos which you can update with `pod repo update` or with `pod install --repo-update`./)
+        e.message.should.match(/ * mistyped the name or version./)
         e.message.should.match(/ * not added the source repo that hosts the Podspec to your Podfile./)
-        e.message.should.match(/ * out-of-date source repos which you can update with `pod repo update`/)
         e.message.should.match(/Note: as of CocoaPods 1.0, `pod repo update` does not happen on `pod install` by default./)
         e.exit_status.should.equal(31)
       end